Move script-style-raw to basic
