+++ packages/integrations/mdx/test/fixtures/basic/src/pages/script-style-raw/index.mdx
